ref(control_silo): Move Email model to Users module
Move email model to the users module and leave shim for getsentry.

ref(# issue here)
