+++ src/sentry/backup/imports.py
@@ -102,10 +102,10 @@ def _import(
     """
 
     # Import here to prevent circular module resolutions.
-    from sentry.models.email import Email
     from sentry.models.organization import Organization
     from sentry.models.organizationmember import OrganizationMember
     from sentry.models.user import User
+    from sentry.users.models.email import Email
 
     if SiloMode.get_current_mode() == SiloMode.CONTROL:
         errText = "Imports must be run in REGION or MONOLITH instances only"
