+++ tests/sentry/backup/test_exports.py
@@ -10,7 +10,6 @@ from sentry.backup.dependencies import NormalizedModelName, get_model, get_model
 from sentry.backup.scopes import ExportScope
 from sentry.backup.validate import validate
 from sentry.db import models
-from sentry.models.email import Email
 from sentry.models.options.option import Option
 from sentry.models.organization import Organization
 from sentry.models.organizationmember import OrganizationMember
@@ -26,6 +25,7 @@ from sentry.testutils.helpers.backups import (
     export_to_file,
 )
 from sentry.testutils.helpers.datetime import freeze_time
+from sentry.users.models.email import Email
 from tests.sentry.backup import get_matching_exportable_models
