+++ tests/sentry/backup/test_imports.py
@@ -35,7 +35,6 @@ from sentry.backup.scopes import ExportScope, ImportScope, RelocationScope
 from sentry.backup.services.import_export.model import RpcImportErrorKind
 from sentry.models.apitoken import DEFAULT_EXPIRATION, ApiToken, generate_token
 from sentry.models.authenticator import Authenticator
-from sentry.models.email import Email
 from sentry.models.importchunk import (
     ControlImportChunk,
     ControlImportChunkReplica,
@@ -81,6 +80,7 @@ from sentry.testutils.helpers.backups import (
 )
 from sentry.testutils.hybrid_cloud import use_split_dbs
 from sentry.testutils.silo import assume_test_silo_mode
+from sentry.users.models.email import Email
 from tests.sentry.backup import (
     expect_models,
     get_matching_exportable_models,
